requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.3
distlib==0.3.4
Django==4.0.2
djangorestframework==3.13.1
filelock==3.4.2
flake8==4.0.1
flake8-import-order==0.18.1
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.4
identify==2.4.8
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.4.1
pre-commit==2.17.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.1
zimports==0.4.4